requirements.txt
langchain
langchain-openai
langgraph
python-dotenv
langchain-core